+++ CHANGELOG.md
@@ -22,6 +22,10 @@ from email
 - mjml files format not to break `trans` instructions
 - configure path for translated files in the backend
 
+### Removed
+
+- Presence activity in the chat
+
 ## [4.0.0-beta.2] - 2022-04-13
 
 ### Added